Refactor: Generalize some metaswap masking logic so that it can be moved to common masking utilities
